// src/domain/enums/nav.enum.ts
export enum NavEnum {
  DASHBOARD = "Dashboard",
  VENTAS = "Ventas",
  PEDIDOS = "Pedidos",
  MENU = "Menu",
  CATEGORIAS = "Categorias",
  PRODUCTOS = "Productos",
  PERSONAL = "Personal",
  PERFILES = "Perfiles",
  USUARIOS = "Usuarios",
  CLIENTES = "Clientes",
  SEDES = "Sedes",
  SETTINGS = "Settings",
  ALERTAS = "Alertas",
}